WIP add callbacks for hide
We added callbacks when the browser is hidden but they won't work because the implementation of SafariViewControllerPlugin is async but doesn't wait for async results. So it notifies us too early when the browser window is closed. Changes to the plugin allow us to open a new window while the old one is closing or properly wait to notify us.